pyproject.toml
[project]
name = "an-resumer"
version = "0.1.0"
description = "Make AN public scrutin more public"
authors = [
    {name = "Dysta"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "dateparser (>=1.2.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "cattrs (>=24.1.2,<25.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
autoflake = "^2.3.1"
isort = "^6.0.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"